rag/requirements.txt
streamlit
chromadb
openai>=1.43.0
python-dotenv
pypdf
